icu_time

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Re-exports

pub use crate::zone::TimeZone;

Modules

iana

🚧 [Unstable] Property names-related data for this component

windows

A provider for mapping Windows Zones to IANA identifiers.

Structs

Baked

Baked data

TimeZoneVariantULE

ULE type for TimeZoneVariant

TimezoneVariantsOffsetsV1

The default mapping between period and offsets. The second level key is a wall-clock time encoded as ZoneNameTimestamp. It represents when the offsets started to be used.