icu_time

Module types

Macros

dt_unit 🔒

This macro defines a struct for 0-based date fields: hours, minutes, seconds and fractional seconds. Each unit is bounded by a range. The traits implemented here will return a Result on whether or not the unit is in range from the given input.

Structs

DateTime

A date and time for a given calendar.

Hour

An ISO-8601 hour component, for use with ISO calendars.

Minute

An ISO-8601 minute component, for use with ISO calendars.

Nanosecond

A fractional second component, stored as nanoseconds.

Second

An ISO-8601 second component, for use with ISO calendars.

Time

A representation of a time in hours, minutes, seconds, and nanoseconds

ZonedDateTime

A date and time for a given calendar, local to a specified time zone.

Constants

UNIX_EPOCH 🔒