icu_time

Struct DateTime

pub struct DateTime<A: AsCalendar> {
    pub date: Date<A>,
    pub time: Time,
}

A date and time for a given calendar.

The primary definition of this type is in the icu_time crate. Other ICU4X crates re-export it for convenience.

Fields
date: Date<A>

The date

time: Time

The time

Implementations

impl<A: AsCalendar> DateTime<A>

pub fn try_from_str(rfc_9557_str: &str, calendar: A) -> Result<Self, ParseError>

Creates a DateTime in any calendar from an RFC 9557 string.

Returns an error if the string has a calendar annotation that does not match the calendar argument, unless the argument is Iso.

โจ Enabled with the ixdtf Cargo feature.

Examples

use icu::calendar::cal::Hebrew;
use icu::time::DateTime;

let datetime =
    DateTime::try_from_str("2024-07-17T16:01:17.045[u-ca=hebrew]", Hebrew)
        .unwrap();

assert_eq!(datetime.date.era_year().year, 5784);
assert_eq!(
    datetime.date.month().standard_code,
    icu::calendar::types::MonthCode(tinystr::tinystr!(4, "M10"))
);
assert_eq!(datetime.date.day_of_month().0, 11);

assert_eq!(datetime.time.hour.number(), 16);
assert_eq!(datetime.time.minute.number(), 1);
assert_eq!(datetime.time.second.number(), 17);
assert_eq!(datetime.time.subsecond.number(), 45000000);

pub fn try_from_utf8( rfc_9557_str: &[u8], calendar: A, ) -> Result<Self, ParseError>

Creates a DateTime in any calendar from an RFC 9557 string.

See Self::try_from_str().

โจ Enabled with the ixdtf Cargo feature.